Refactor content-css as basic fixture
