--- packages/integrations/mdx/test/fixtures/astro-content-css/src/content/config.ts
@@ -1,12 +1,0 @@
-// 1. Import utilities from `astro:content`
-import { z, defineCollection } from 'astro:content';
-// 2. Define a schema for each collection you'd like to validate.
-const dynamicCollection = defineCollection({
-  schema: z.object({
-    title: z.string(),
-  }),
-});
-// 3. Export a single `collections` object to register your collection(s)
-export const collections = {
-  dynamic: dynamicCollection,
-};
