+++ packages/integrations/mdx/test/fixtures/basic/astro.config.mjs
@@ -1,11 +1,6 @@
 import { defineConfig } from 'astro/config';
-
 import mdx from "@astrojs/mdx";
 
-// https://astro.build/config
 export default defineConfig({
-  build: {
-    format: 'file'
-  },
   integrations: [mdx()]
 });
