+++ packages/integrations/mdx/test/fixtures/basic/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@test/astro-content-css",
+  "name": "@test/mdx-basic",
   "version": "0.0.0",
   "private": true,
   "dependencies": {
