+++ packages/integrations/mdx/test/fixtures/basic/src/content/astro-content-css/FirstComponentWithJS.astro
