+++ packages/integrations/mdx/test/fixtures/basic/src/content/astro-content-css/first-component-with-js.mdx
