+++ packages/integrations/mdx/test/fixtures/basic/src/content/config.ts
@@ -1,0 +1,11 @@
+import { z, defineCollection } from 'astro:content';
+
+const astroContentCssCollection = defineCollection({
+  schema: z.object({
+    title: z.string(),
+  }),
+});
+
+export const collections = {
+  ['astro-content-css']: astroContentCssCollection,
+};
