+++ packages/integrations/mdx/test/fixtures/basic/src/pages/astro-content-css/index.astro
@@ -1,7 +1,7 @@
 ---
 import { getCollection } from 'astro:content';
 
-const entries = await getCollection('dynamic');
+const entries = await getCollection('astro-content-css');
 ---
 
 <!DOCTYPE html>
